Strict Null Checks - background/contextMenus (#8387)
* separate ensureContextMenu

* fix imports

* moves rest of exported backgroudn contextMenu functions to separate files

* fix tests
