+++ src/background/messenger/strict/api.ts
@@ -123,3 +123,9 @@ export const collectPerformanceDiagnostics = getMethod(
 
 export const initTelemetry = getNotifier("INIT_TELEMETRY", bg);
 export const sendDeploymentAlert = getNotifier("SEND_DEPLOYMENT_ALERT", bg);
+
+export const ensureContextMenu = getMethod("ENSURE_CONTEXT_MENU", bg);
+/**
+ * Uninstall context menu and return whether the context menu was uninstalled.
+ */
+export const uninstallContextMenu = getMethod("UNINSTALL_CONTEXT_MENU", bg);
