+++ src/background/messenger/strict/registration.ts
@@ -64,6 +64,8 @@ import {
   recordEvent,
   sendDeploymentAlert,
 } from "@/background/telemetry";
+import { ensureContextMenu } from "@/background/contextMenus/ensureContextMenu";
+import { uninstallContextMenu } from "@/background/contextMenus/uninstallContextMenu";
 
 expectContext("background");
 
@@ -123,6 +125,9 @@ declare global {
     RECORD_EVENT: typeof recordEvent;
     INIT_TELEMETRY: typeof initTelemetry;
     SEND_DEPLOYMENT_ALERT: typeof sendDeploymentAlert;
+
+    ENSURE_CONTEXT_MENU: typeof ensureContextMenu;
+    UNINSTALL_CONTEXT_MENU: typeof uninstallContextMenu;
   }
 }
 
@@ -182,5 +187,8 @@ export default function registerMessenger(): void {
     RECORD_EVENT: recordEvent,
     INIT_TELEMETRY: initTelemetry,
     SEND_DEPLOYMENT_ALERT: sendDeploymentAlert,
+
+    ENSURE_CONTEXT_MENU: ensureContextMenu,
+    UNINSTALL_CONTEXT_MENU: uninstallContextMenu,
   });
 }
