+++ src/starterBricks/contextMenu/contextMenu.test.ts
@@ -29,7 +29,7 @@ import { RunReason } from "@/types/runtimeTypes";
 import {
   uninstallContextMenu,
   ensureContextMenu,
-} from "@/background/messenger/api";
+} from "@/background/messenger/strict/api";
 import { getPlatform } from "@/platform/platformContext";
 import {
   type ContextMenuDefinition,
