+++ src/tsconfig.strictNullChecks.json
@@ -76,6 +76,9 @@
     "./background/capture.ts",
     "./background/clipboard.ts",
     "./background/contentScript.ts",
+    "./background/contextMenus/ensureContextMenu.ts",
+    "./background/contextMenus/makeMenuId.ts",
+    "./background/contextMenus/uninstallContextMenu.ts",
     "./background/dataStore.ts",
     "./background/externalProtocol.ts",
     "./background/initTheme.ts",
